chore: update content-type for provisioning to 'application/json'
